# WEBSITE AI CLOUD/SWARM-AI-Website-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SWARM AI chat, provider settings, and health API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: List provider connection status without revealing credentials
      responses:
        "200":
          description: Provider settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /providers/{providerId}:
    parameters:
      - name: providerId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/ProviderId"
    put:
      operationId: saveProvider
      tags: [providers]
      summary: Save a provider key, encrypted for the signed-in user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveProviderInput"
      responses:
        "200":
          description: Saved provider status, with only a masked key hint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteProvider
      tags: [providers]
      summary: Remove the signed-in user's provider key
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required: [error]
    ProviderId:
      type: string
      enum: [openai, openrouter, nvidia, groq, google, cloudflare, huggingface, cerebras, mistral]
...
    SaveProviderInput:
      type: object
      properties:
        apiKey:
          type: string
          minLength: 8
          maxLength: 4096
        accountId:
          type: string
          maxLength: 160
      required: [apiKey]
    ProviderSummary:
      type: object
      properties:
        providerId:
          $ref: "#/components/schemas/ProviderId"
        name:
          type: string
        configured:
          type: boolean
        needsAccountId:
          type: boolean
        accountId:
          type: string
          nullable: true
        keyHint:
          type: string
          nullable: true
      required: [providerId, name, configured, needsAccountId, accountId, keyHint]
    ProviderListResponse:
      type: object
      properties:
        providers:
          type: array
          items:
            $ref: "#/components/schemas/ProviderSummary"
      required: [providers]
    ProviderResponse:
      type: object
      properties:
        provider:
          $ref: "#/components/schemas/ProviderSummary"
      required: [provider]
...