# lib/cloud-api/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SWARM AI chat, provider settings, and health API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}

export type ProviderId = typeof ProviderId[keyof typeof ProviderId];


export const ProviderId = {
  openai: 'openai',
  openrouter: 'openrouter',
  nvidia: 'nvidia',
  groq: 'groq',
  google: 'google',
  cloudflare: 'cloudflare',
  huggingface: 'huggingface',
  cerebras: 'cerebras',
  mistral: 'mistral',
} as const;

export interface ConversationSummary {
  id: string;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export type Conversation = ConversationSummary & {
  messages: Message[];
};

export interface ConversationListResponse {
  conversations: ConversationSummary[];
}

export interface CreateConversationInput {
  /** @maxLength 120 */
  title?: string;
}

export interface RenameConversationInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
}

export type StreamConversationInputAction = typeof StreamConversationInputAction[keyof typeof StreamConversationInputAction];


export const StreamConversationInputAction = {
  send: 'send',
  edit: 'edit',
  regenerate: 'regenerate',
  continue: 'continue',
} as const;

export type StreamConversationInputMode = typeof StreamConversationInputMode[keyof typeof StreamConversationInputMode];


export const StreamConversationInputMode = {
  chat: 'chat',
  swarm: 'swarm',
} as const;

export interface StreamConversationInput {
  action: StreamConversationInputAction;
  /** @maxLength 32000 */
  prompt?: string;
  messageId?: number;
  providerId: ProviderId;
  /**
     * @minLength 1
     * @maxLength 200
     */
  modelId: string;
  mode: StreamConversationInputMode;
}

export interface SaveProviderInput {
  /**
     * @minLength 8
     * @maxLength 4096
     */
  apiKey: string;
  /** @maxLength 160 */
  accountId?: string;
}

export interface ProviderSummary {
  providerId: ProviderId;
  name: string;
  configured: boolean;
  needsAccountId: boolean;
  accountId: string | null;
  keyHint: string | null;
}

export interface ProviderListResponse {
  providers: ProviderSummary[];
}

export interface ProviderResponse {
  provider: ProviderSummary;
}

export interface ModelOption {
  id: string;
  modelId: string;
  providerId: ProviderId;
  providerName: string;
  displayName: string;
}

export interface ModelListResponse {
  models: ModelOption[];
}

/**
 * Authentication required
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

export type ListConversationsParams = {
q?: string;
};